update rest Cognapps to the scheme with RuleURI as input+ implantator for SemFormToRule + update of all inputs/outputs
Similar to the previous commit we have adapted also the SemFormToRule
and SemFormToPatient Cognitive Apps to the scheme where the input is a
ruleURI, not a file. This ruleURI equals actually to the Wiki page URI.
Added an Implantator for SemFormToRule which names the start node of the
appropriate swrl rule to the ruleName#1 (trying to achieve the same
structure as by guideline based rules). Further input / output example
patterns for all cognitive apps have been adapted.